# hospital-management-system/appointment-service/src/main/java/com/hospital/appointment_service/entities/Appointment.java
package com.hospital.appointment_service.entities;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import java.time.LocalDateTime;

@Entity
public class Appointment {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private LocalDateTime date;
    private Long patientId;

    public Appointment() {}

    public Appointment(Long id, LocalDateTime date, Long patientId) {
        this.id = id;
        this.date = date;
        this.patientId = patientId;
    }

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public LocalDateTime getDate() { return date; }
    public void setDate(LocalDateTime date) { this.date = date; }
    public Long getPatientId() { return patientId; }
    public void setPatientId(Long patientId) { this.patientId = patientId; }
}