# hospital-management-system/medical-record-service/src/main/java/com/hospital/medical_record_service/entities/MedicalRecord.java
package com.hospital.medical_record_service.entities;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import java.time.LocalDateTime;

@Entity
public class MedicalRecord {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private Long patientId;
    private String diagnostics;
    private LocalDateTime createdAt;

    public MedicalRecord() {
        this.createdAt = LocalDateTime.now();
    }

    public MedicalRecord(Long id, Long patientId, String diagnostics) {
        this.id = id;
        this.patientId = patientId;
        this.diagnostics = diagnostics;
        this.createdAt = LocalDateTime.now();
    }

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public Long getPatientId() { return patientId; }
    public void setPatientId(Long patientId) { this.patientId = patientId; }
    public String getDiagnostics() { return diagnostics; }
    public void setDiagnostics(String diagnostics) { this.diagnostics = diagnostics; }
    public LocalDateTime getCreatedAt() { return createdAt; }
    public void setCreatedAt(LocalDateTime createdAt) { this.createdAt = createdAt; }
}